requirements.txt
pyrofork
pymongo
yt-dlp
tgcrypto
aiofiles
requests
python-dotenv
motor
hachoir
pillow
